\name{names}
\alias{names.ScriptEngine}
\alias{names.CompiledScript}

\docType{methods}

\title{
Retrieve the Names for a ScriptEngine or CompiledScript class
}

\description{
These methods retrieve the names associated with objects of class ScriptEngine or CompiledScript.
}

\usage{
  \method{names}{ScriptEngine}(x, ...)
  \method{names}{CompiledScript}(x, ...)
}

\arguments{
  \item{x}{An object of class ScriptEngine or CompiledScript.}
  \item{...}{Ignored.}
}

\keyword{ interface }

