\name{print}
\alias{print.ScriptEngine}
\alias{toString.ScriptEngine}
\alias{print.CompiledScript}
\alias{toString.CompiledScript}

\docType{methods}

\title{
Print or Return a Character Representation of a Script Engine or Compiled Script Class Object
}

\description{
These methods print or return a character representation of a script engine.
}

\usage{
  \method{print}{ScriptEngine}(x, ...)
  \method{toString}{ScriptEngine}(x, ...)
  \method{print}{CompiledScript}(x, ...)
  \method{toString}{CompiledScript}(x, ...)
}

\arguments{
  \item{x}{An object of class ScriptEngine or CompiledScript.}
  \item{...}{Currently ignored.}
}

\keyword{ interface }

